% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen2 fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Function "fars_map_state" - Display accidents map by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer with the State Code}

\item{year}{a string or an integer}
}
\value{
This function returns a map with the accidents in the speicified state at
the specified year. It returns NULL and warning message if no accidents recorded
and returns warning message if the state number is invalid.
}
\description{
This function displays a plot with a state map including the accidents
location by year. The functions \link{make_filename} and \link{fars_read}
contribute to this function. If the \code{state.num} is invalid the function shows
an error message "invalid STATE number: ". If there are no accidents in the speicific
state in the specified year the function shows warning message "no accidents to plot".
}
\examples{
\dontrun{
fars_map_state(42, 2013)
}
}
\seealso{
\link{fars_read}

\link{make_filename}
}
