% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen2 fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{FARS functions

These functions provided are using data from the US National Highway Traffic
Safety Administration's Fatality Analysis Reporting System (FARS), which is
a nationwide census providing the American public yearly data regarding
fatal injuries suffered in motor vehicle traffic crashes.


Function "fars_read" - Open CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the file name with its location.
This parameter generated by: \code{\link{fars_map_state}}}
}
\value{
This function returns a data frame that containes data from CSV file.
If file does not exist it returns an error message.
}
\description{
This function reads CSV file and opens it in a data frame form.
}
\details{
This function opens CSV file from FARS. If the file with the specified name
does not exist the function gives an error message"file....does not exist".
Function suppress any warning messages that can be generated by read_csv function
used to open CSV file. It contributes to the function \link{fars_read_years}.
}
\examples{
\dontrun{
library(dplyr)
library(readr)
library(magrittr)
library(tidyr)
library(dplyr)
library(maps)
library(graphics)

fars_read("accident_2015.csv.bz2")}


}
