% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function "fars_summarize_years" - Summarize data from FARS}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector with the list of years}
}
\value{
This function returns a data frame with number of accidents by years
summarized by month
}
\description{
. This function summarizes yearly accidents data, by month. it uses the
data frames created by the function fars_read_years. It merges data frames,
groups data by year and month and spread data by columns
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)}

}
