% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Function "make_filename" - creates a file name in a specified format}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or an integer}
}
\value{
This function returns a string with the CSV file name for a given
 year
}
\description{
Function creates a string to be used as a name of CSV file
using a given year and a string "accident_" ... ".csv.bz2".
}
\details{
This function contributes to the functions fars_read_years,
fars_map_state
}
\note{
The returned string of this function used in the function fars_read
and as an input in the function fars_map_state to make an input parameter
filename
}
\examples{
\dontrun{
make_filename(2015)}


}
