% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen2 fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Function "make_filename" - creates a file name in a specified format}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or an integer}
}
\value{
This function returns a string with the CSV file name for a given
 year
}
\description{
Function creates a string to be used as a name of CSV file
using given \code{year} and a string "accident_" ... ".csv.bz2".
}
\details{
This function contributes to the functions \link{fars_read_years},
\link{fars_map_state}
}
\examples{
\dontrun{
make_filename(2015)}

}
\seealso{
\link{fars_read}, \link{fars_map_state}
}
